refactor(filter-bar): replace setFilterActive with updateSingleFilter
