+++ packages/components/src/FilterBar/context/reducer/filtersStateReducer.ts
@@ -1,5 +1,4 @@
 import { FilterState, FiltersState, FiltersValues } from "../types"
-import { setFilterActive } from "./setFilterActive"
 import { updateSingleFilter } from "./updateSingleFilter"
 import { updateValues } from "./updateValues"
 
@@ -8,9 +7,8 @@ type Actions<ValuesMap> =
   | {
       type: "update_single_filter"
       id: keyof ValuesMap
-      data: Partial<FilterState<keyof ValuesMap, ValuesMap>>
+      data: Partial<FilterState<keyof ValuesMap, ValuesMap[keyof ValuesMap]>>
     }
-  | { type: "set_filter_active"; id: keyof ValuesMap; value: boolean }
 
 export const filtersStateReducer = <ValuesMap extends FiltersValues>(
   state: FiltersState<ValuesMap>,
@@ -22,8 +20,5 @@ export const filtersStateReducer = <ValuesMap extends FiltersValues>(
 
     case "update_single_filter":
       return updateSingleFilter(state, action.id, action.data)
-
-    case "set_filter_active":
-      return setFilterActive(state, action.id, action.value)
   }
 }
